+++ source/YoutubeDLSharp/Options/MultiValue.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 
 namespace YoutubeDLSharp.Options;
 
@@ -8,24 +7,11 @@ public sealed class MultiValue<T>
 {
     public List<T> Values { get; }
 
-    public MultiValue(params T[] values)
+    public MultiValue(params ReadOnlySpan<T> values)
     {
-        Values = values.ToList();
+        Values = [];
+        Values.AddRange(values);
     }
 
     public static implicit operator MultiValue<T>(T value) => new(value);
-
-    public static implicit operator MultiValue<T>(T[] values) => new(values);
-
-    public static explicit operator T(MultiValue<T> value)
-    {
-        if (value.Values is [var singleValue])
-        {
-            return singleValue;
-        }
-
-        throw new InvalidCastException($"Cannot cast sequence of values to {typeof(T)}.");
-    }
-
-    public static explicit operator T[](MultiValue<T> value) => value.Values.ToArray();
 }